refactor: payment DTO를 controller 패키지로 이동
- payment/domain/dto/* → payment/controller/dto/* 이동
- 웹 계층과 도메인 계층 책임 분리
