+++ src/main/java/com/ururulab/ururu/payment/controller/dto/response/PaymentConfirmResponseDto.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.payment.domain.dto.response;
+package com.ururulab.ururu.payment.controller.dto.response;
 
 import com.ururulab.ururu.payment.domain.entity.enumerated.PaymentStatus;
