+++ src/main/java/com/ururulab/ururu/payment/controller/dto/response/PaymentFailResponseDto.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.payment.domain.dto.response;
+package com.ururulab.ururu.payment.controller.dto.response;
 
 public record PaymentFailResponseDto(
         String errorCode,
